backend/requirements.txt
Django==4.1.7
django-filter==23.2
djangorestframework==3.14.0
djoser==2.1.0
drf-extra-fields==3.4.1
Faker==18.7.0
Pillow==9.4.0
urllib3==2.0.2
gunicorn==20.0.4
psycopg2-binary==2.9.3
python-dotenv==1.0.0
